Add `app.command.ToggleWorkspaceLayout` to the command documentation
As far as I could see it does not include any parameters. I'm unsure about the new UI markers and whether they would be required here.

I also just noticed this feature is still in beta, which might be why it wasn't in the documentation yet. But since I had already written it down anyway, might as well offer it up!